moved Bubbles to a component folders
