+++ src/components/bubbles/SpeechBubble.jsx
@@ -2,17 +2,13 @@ import React, { useEffect } from 'react'
 import './Bubble.css'
 import { useTranslation } from 'react-i18next'
 import { useSelector, useDispatch } from 'react-redux';
-import { initializePageAttributes, selectPageAttributes } from '../store/slices/pageSlice';
-
-/**
- * Like SpeechBubble but it doesn't bounce and glow.
- */
+import { initializePageAttributes, selectPageAttributes } from '../../store/slices/pageSlice';
 
 const defaultAttributes = {
   show: false
 }
 
-const NarrativeBubble = ({ pageId="", id="", mainText, subText, characterSrc, showNext=false, onClick, top = '-50%',bottom, left = '0vh', maxWidth="300px", right, style }) => {
+const SpeechBubble = ({ pageId="", id="", mainText, subText, characterSrc, showNext=false, onClick, top = '-50%',bottom, left = '0vh', maxWidth="300px", right, style }) => {
   const { t } = useTranslation()
   const dispatch = useDispatch();
   const { show } = useSelector(state => selectPageAttributes(state, `${pageId}:${id}`));
@@ -28,7 +24,7 @@ const NarrativeBubble = ({ pageId="", id="", mainText, subText, characterSrc, sh
     maxWidth
   }
   return show ? (
-    <div className="narrative-bubble bubble" style={bubbleStyle} onClick={onClick}>
+    <div className="speech-bubble bubble" style={bubbleStyle} onClick={onClick}>
       {characterSrc && <img src={characterSrc} style={{position: "absolute", left: "-40%", top: "-50%", width: "15vh"}} />}
       <span style={{ display: 'block' }}>{mainText}</span>
       {subText && (
@@ -58,4 +54,4 @@ const NarrativeBubble = ({ pageId="", id="", mainText, subText, characterSrc, sh
   ) : null
 }
 
-export default NarrativeBubble
+export default SpeechBubble
