+++ src/helpers/bubbleHelper.js
@@ -1,5 +1,5 @@
 
-import { setPageAttribute, selectPageAttributes } from '../store/slices/pageSlice';
+import { setPageAttribute, selectPageAttributes } from '../../store/slices/pageSlice';
 
 export const setBubbleShow = (pageId, bubbleId, show) => setPageAttribute({pageId: `${pageId}:${bubbleId}`, key: 'show', value: show })
 export const selectBubbleShowAttribute = (state, pageId, bubbleId, defaultValue) => { return selectPageAttributes(state, `${pageId}:${bubbleId}`, {'show':defaultValue})['show'] }
